requirements.txt
Django
bootstrap4
django-bootstrap4
django-bootstrap-datepicker-plus